+++ front-end-project-for-backend-week/lambdaNotes/serverDerp/clients/userModel.js
@@ -40,4 +40,6 @@ UserSchema.methods.checkPassword = function(plainTextPW, callBack) {
     })
 };
 
-module.exports = mongoose.model('User', UserSchema);
+const User = mongoose.model('User', UserSchema);
+
+module.exports = User;
